+++ src/main/kotlin/me/nobaboy/nobaaddons/commands/internal/CommandUtil.kt
@@ -2,7 +2,6 @@ package me.nobaboy.nobaaddons.commands.internal
 
 import com.mojang.brigadier.CommandDispatcher
 import com.mojang.brigadier.builder.LiteralArgumentBuilder
-import net.fabricmc.fabric.api.client.command.v2.ClientCommandManager
 import net.fabricmc.fabric.api.client.command.v2.ClientCommandRegistrationCallback
 import net.fabricmc.fabric.api.client.command.v2.FabricClientCommandSource
 
@@ -16,12 +15,13 @@ object CommandUtil {
 	}
 
 	fun register(root: ICommand, dispatcher: CommandDispatcher<FabricClientCommandSource>) {
+		if(!root.enabled) return
 		val names = listOf(root.name, *root.aliases.toTypedArray())
 		names.forEach { dispatcher.register(root.create(it)) }
 	}
 
 	fun addAll(command: LiteralArgumentBuilder<FabricClientCommandSource>, commands: List<ICommand>) {
-		commands.forEach {
+		commands.filter { it.enabled }.forEach {
 			val names = listOf(it.name, *it.aliases.toTypedArray())
 			names.forEach { name -> command.then(it.create(name)) }
 		}