+++ src/main/kotlin/me/nobaboy/nobaaddons/events/SoundEvents.kt
@@ -4,6 +4,7 @@ import me.nobaboy.nobaaddons.events.internal.CancelableEvent
 import me.nobaboy.nobaaddons.events.internal.CancelableEventDispatcher
 import me.nobaboy.nobaaddons.events.internal.EventDispatcher
 import me.nobaboy.nobaaddons.utils.NobaVec
+import net.minecraft.sound.SoundCategory
 import net.minecraft.util.Identifier
 
 object SoundEvents {
@@ -13,6 +14,9 @@ object SoundEvents {
 	@JvmField
 	val SOUND = EventDispatcher<Sound>()
 
+	@JvmField
+	val SOUND_CANCELED = EventDispatcher<Sound>()
+
 	data class AllowSound(val id: Identifier, val location: NobaVec, val pitch: Float, val volume: Float) : CancelableEvent()
-	data class Sound(val id: Identifier, val location: NobaVec, val pitch: Float, val volume: Float)
+	data class Sound(val id: Identifier, val category: SoundCategory, val location: NobaVec, val pitch: Float, val volume: Float)
 }